requirement.txt
pytorch-lightning==1.5.0
einops
opencv-python==4.7.0.72
matplotlib
omegaconf
transformers==4.33.2
xformers==0.0.19
triton==2.0.0
open-clip-torch==2.19.0
diffusers==0.20.2
scipy==1.10.1
tqdm